AP_ADSB: correct bad 16bit port read in ADSB drivers

+++ libraries/AP_ADSB/AP_ADSB_Sagetech.cpp
@@ -73,7 +73,7 @@ void AP_ADSB_Sagetech::update()
         if (!_port->read(data)) {
             break;
         }
-        if (parse_byte_XP((uint8_t)data)) {
+        if (parse_byte_XP(data)) {
             handle_packet_XP(message_in.packet);
         }
     } // while nbytes
